requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
vital==2.1.219